requirements.txt
streamlit>=1.30.0
opencv-python>=4.8.0
ultralytics
pandas>=2.0.0
numpy>=2.1.0
Pillow>=10.0.0
torch
torchvision
facenet-pytorch

